Fix StoryView preview
